requirements.txt
discord.py
git+https://github.com/bravosierra99/dpytest.git
python-dotenv
fake_useragent
lxml
nekos.py
Pillow
psycopg2; sys_platform == "win32"
psycopg2-binary; sys_platform == "linux"
pybooru
python-chess
requests
